Extracts Java forms to play-java-forms project
Had to fix ResourcesTest because the complete async happened after the exception was returned.
