+++ framework/src/play-java-forms/src/main/resources/reference.conf
@@ -1,0 +1,7 @@
+play {
+  modules {
+    enabled += "play.data.FormFactoryModule"
+    enabled += "play.data.format.FormattersModule"
+    enabled += "play.data.validation.ValidatorsModule"
+  }
+}
