+++ framework/src/play-java-forms/src/test/scala/play/data/DynamicFormSpec.scala
@@ -9,7 +9,7 @@ import play.api.{ Configuration, Environment }
 import play.api.i18n.{ DefaultLangs, DefaultMessagesApi }
 import play.data.format.Formatters
 import views.html.helper.inputText
-import play.core.j.PlayMagicForJava.javaFieldtoScalaField
+import play.core.j.PlayFormsMagicForJava.javaFieldtoScalaField
 import views.html.helper.FieldConstructor.defaultField
 import scala.collection.JavaConversions._
 import javax.validation.Validation
