+++ framework/src/play-java/src/main/java/play/inject/BuiltInModule.java
@@ -6,16 +6,11 @@ package play.inject;
 import play.api.Configuration;
 import play.api.Environment;
 import play.api.inject.Binding;
-import play.data.validation.DefaultConstraintValidatorFactory;
-import play.data.validation.ValidatorProvider;
 import play.libs.crypto.CSRFTokenSigner;
 import play.libs.crypto.CookieSigner;
 import play.libs.crypto.DefaultCSRFTokenSigner;
 import play.libs.crypto.HMACSHA1CookieSigner;
 import scala.collection.Seq;
-
-import javax.validation.ConstraintValidatorFactory;
-import javax.validation.Validator;
 
 public class BuiltInModule extends play.api.inject.Module {
     @Override
@@ -23,8 +18,6 @@ public class BuiltInModule extends play.api.inject.Module {
         return seq(
           bind(ApplicationLifecycle.class).to(DelegateApplicationLifecycle.class),
           bind(play.Configuration.class).toProvider(ConfigurationProvider.class),
-          bind(ConstraintValidatorFactory.class).to(DefaultConstraintValidatorFactory.class),
-          bind(Validator.class).toProvider(ValidatorProvider.class),
           bind(CSRFTokenSigner.class).to(DefaultCSRFTokenSigner.class),
           bind(CookieSigner.class).to(HMACSHA1CookieSigner.class)
         );
