+++ framework/src/play-java/src/main/resources/reference.conf
@@ -1,8 +1,6 @@
 play {
   modules {
     enabled += "play.inject.BuiltInModule"
-    enabled += "play.data.FormFactoryModule"
-    enabled += "play.data.format.FormattersModule"
     enabled += "play.core.ObjectMapperModule"
   }
 }
