+++ framework/src/play-java/src/main/scala/play/core/TemplateMagicForJava.scala
@@ -9,7 +9,6 @@ import scala.util.control.NonFatal
 /** Defines a magic helper for Play templates in a Java context. */
 object PlayMagicForJava {
 
-  import scala.collection.JavaConverters._
   import scala.language.implicitConversions
   import scala.compat.java8.OptionConverters._
 
@@ -21,35 +20,6 @@ object PlayMagicForJava {
       play.mvc.Http.Context.Implicit.lang.asInstanceOf[play.api.i18n.Lang]
     } catch {
       case NonFatal(_) => play.api.i18n.Lang.defaultLang
-    }
-  }
-
-  /**
-   * Implicit conversion of a Play Java form `Field` to a proper Scala form `Field`.
-   */
-  implicit def javaFieldtoScalaField(jField: play.data.Form.Field): play.api.data.Field = {
-
-    new play.api.data.Field(
-      null,
-      jField.name,
-      Option(jField.constraints).map(c => c.asScala.map { jT =>
-        jT._1 -> jT._2.asScala
-      }).getOrElse(Nil),
-      Option(jField.format).map(f => f._1 -> f._2.asScala),
-      Option(jField.errors).map(e => e.asScala.map { jE =>
-        play.api.data.FormError(
-          jE.key,
-          jE.messages.asScala,
-          jE.arguments.asScala)
-      }).getOrElse(Nil),
-      Option(jField.value)) {
-
-      override def apply(key: String) = {
-        javaFieldtoScalaField(jField.sub(key))
-      }
-
-      override lazy val indexes = jField.indexes.asScala.toSeq.map(_.toInt)
-
     }
   }
